pyproject.toml
[project]
name = "lilbee"
version = "0.6.0"
description = "Local knowledge base for documents and code. Search, ask questions, or chat — standalone or as an AI agent backend via MCP. Runs fully offline with llama-cpp."
readme = "README.md"
license = "MIT"
authors = [{ name = "tobocop2", email = "5562156+tobocop2@users.noreply.github.com" }]
requires-python = ">=3.11"
keywords = [
    "knowledge-base", "rag", "lancedb", "mcp",
    "vector-search", "ocr", "pdf", "local-llm", "offline",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
    "Typing :: Typed",
]

dependencies = [
    "lancedb",
    "kreuzberg>=4.6.3",
    "filelock",
    "tree-sitter-language-pack>=1.4.0",
    "typer",
    "tiktoken",
    "textual>=0.75",
    "mcp>=1.26.0",
    "pillow>=11.3.0",
    "litestar>=2.0",
    "uvicorn>=0.30",
    "httpx",
    "llama-cpp-python",
    "huggingface_hub",  # TODO: remove path override after PR merges
    # "huggingface_hub",
    "psutil>=5.9",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
litellm = ["litellm"]
reranker = ["sentence-transformers"]
graph = ["spacy>=3.8", "graspologic-native>=1.2"]
crawler = ["crawl4ai>=0.8.0"]

[project.urls]
Homepage = "https://tobocop2.github.io/lilbee/"
Repository = "https://github.com/tobocop2/lilbee"
Issues = "https://github.com/tobocop2/lilbee/issues"

[project.scripts]
lilbee = "lilbee.launcher:main"

[tool.uv.sources]
huggingface_hub = { path = "../huggingface_hub", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lilbee"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pexpect>=4.9.0",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "reportlab",
    "ruff>=0.15.4",
    "httpx>=0.27",
    "pytest-httpserver",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = [".", "tests"]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--ignore=tests/integration"
markers = ["slow: integration tests that download models or take >10s"]

[tool.coverage.run]
source = ["lilbee"]
omit = ["src/lilbee/__main__.py", "src/lilbee/launcher.py"]
concurrency = ["thread"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if sys.platform",
    "class.*Protocol",
]
omit = [
    "src/lilbee/providers/litellm_provider.py",
    "src/lilbee/__main__.py",
    "src/lilbee/launcher.py",
    "src/lilbee/platform.py",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests", "tools"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
